test(io): move utility function to util.py
The function `match_log` should be in `util.py` so it can be reused
in other modules.
